feat(buzz): the provider moves to the extension, and actually ships (#1508)
ADR 0043 gate 5. `buzz-backend-fountain` moves to the extension's ownership;
`fountain buzz` stays in `cli/`.

## The rescope

The issue proposed moving both Go surfaces. ADR 0043 decision 6 settles it the
other way for one of them, and the issue was rescoped to match:

`fountain buzz agents ...` is a thin HTTP client against `/api/buzz/agents`
with no Nostr code and no server state. Subtracting a subcommand from a
released binary is a user-visible break with nothing bought, so it stays,
covered by `cli/internal/cmd/docs_test.go`'s diff against `docs/cli.md`.

`buzz-backend-fountain` is a separate `main` package that Buzz's desktop
discovers by name on `PATH`, never through `fountain`, so it moves with no
compatibility question at all. That is one executable, not two, and there is
no new `fountain-buzz` management binary.

## A module of its own, so the compiler holds the boundary

`apps/fountain_buzz/cli` is a Go module, not just a directory. Go's `internal/`
rule is scoped to a module path, so the provider cannot reach into
`cli/internal/...` even by accident — the same boundary
`Fountain.ExtensionGuardTest` enforces on the Elixir side, enforced here by the
compiler instead of a guard test. Graduation (#1510) then becomes a directory
move plus turning one `replace` into a version pin.

That constraint is the reason `cli/internal/{api,config,credentials}` are now
`cli/{api,config,credentials}`: the provider needs `New`, `Get`, `Post` and
ambient credential resolution, and an out-of-module consumer cannot import an
`internal` package. Twenty files' imports rewritten, no behaviour touched.

## The half that was never built

The issue's gate says "release artifacts publish both plugin executables",
which reads like a property to preserve. It was not true for either.
`release.yml` built exactly `./cmd/fountain`, and `buzz-backend-fountain`
appeared in no workflow, no Dockerfile stage and no install path:

    $ gh release view v0.16.0 --json assets --jq '.assets[].name'
    fountain-darwin-amd64
    fountain-darwin-arm64
    fountain-linux-amd64
    fountain-linux-arm64
    openapi.json

It was compiled only by `go test ./...` and shipped to nobody, while
`docs/integrations/buzz.md` said "Fountain ships a Buzz remote-agents provider"
and ADR 0043 said it had "its own release artifact". That is the
describes-unbuilt-behavior-as-existing failure this repository has a rule
against, and a reader of either concluded the desktop provider path worked.

It now cross-compiles for the same four platforms as the `fountain` binary and
attaches to every release, so both documents are true rather than aspirational.
The docs gained install instructions, and say plainly that a core distribution
serves no Buzz routes and the deploy gets a 404.

The provider is not in any image: it runs on the desktop machine beside Buzz,
not on the server. The Homebrew tap job filters artifacts on `fountain-*` and
is unaffected.

## Verified

- Both modules: `go build`, `go vet`, `gofmt` and `go test -count=1` clean.
- Cross-compiles for linux/darwin x amd64/arm64.
- The built binary answers the `info` op with `protocol_version 1` and its
  config schema — ADR 0043 decision 6's compatibility promise, end to end.
- `docs_test.exs`, `docs-style.py`, `destink` and `vale lint docs` clean;
  `okf validate decisions` valid.

`-count=1` on the new steps matches what #1543 established, and the CLI docs
parity step gets it too: those tests read `docs/cli.md` from outside their
package and run precisely when the docs changed and `cli/` may not have, which
is #1542's hazard exactly.
