+++ .github/workflows/release.yml
@@ -33,15 +33,19 @@ jobs:
           - goos: linux
             goarch: amd64
             artifact: fountain-linux-amd64
+            provider_artifact: buzz-backend-fountain-linux-amd64
           - goos: linux
             goarch: arm64
             artifact: fountain-linux-arm64
+            provider_artifact: buzz-backend-fountain-linux-arm64
           - goos: darwin
             goarch: amd64
             artifact: fountain-darwin-amd64
+            provider_artifact: buzz-backend-fountain-darwin-amd64
           - goos: darwin
             goarch: arm64
             artifact: fountain-darwin-arm64
+            provider_artifact: buzz-backend-fountain-darwin-arm64
 
     steps:
       - uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v4
@@ -63,6 +67,13 @@ jobs:
         working-directory: cli
         run: go test -count=1 ./...
 
+      # The Buzz provider is a second module and `go test ./...` above does not
+      # reach it. Same reason as the step above: a release used to
+      # cross-compile without running the suite.
+      - name: Test the Buzz provider
+        working-directory: apps/fountain_buzz/cli
+        run: go test -count=1 ./...
+
       - name: Build
         env:
           CGO_ENABLED: "0"
@@ -82,11 +93,42 @@ jobs:
           file "../dist/${{ matrix.artifact }}"
           du -h "../dist/${{ matrix.artifact }}"
 
+      # `buzz-backend-fountain` is the Buzz remote-agents provider. Buzz's
+      # desktop discovers it by name on PATH, so it has to BE somewhere a
+      # person can put on their PATH — and until #1508 it was built by nothing
+      # and attached to no release, while docs/integrations/buzz.md and ADR
+      # 0043 both described it as shipped. It is a separate module owned by
+      # the extension, cross-compiled for the same four platforms as the
+      # `fountain` binary beside it.
+      #
+      # No version stamp: the provider reports `protocol_version` (1) and its
+      # own `version` string in the `info` op, which is what Buzz's desktop
+      # checks. Neither is the Fountain release tag.
+      - name: Build the Buzz provider
+        env:
+          CGO_ENABLED: "0"
+          GOOS: ${{ matrix.goos }}
+          GOARCH: ${{ matrix.goarch }}
+        working-directory: apps/fountain_buzz/cli
+        run: |
+          set -euxo pipefail
+          mkdir -p ../../../dist
+          go build -trimpath -ldflags "-s -w" \
+            -o "../../../dist/${{ matrix.provider_artifact }}" ./cmd/buzz-backend-fountain
+          file "../../../dist/${{ matrix.provider_artifact }}"
+          du -h "../../../dist/${{ matrix.provider_artifact }}"
+
       - name: Upload artifact
         uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
         with:
           name: ${{ matrix.artifact }}
           path: dist/${{ matrix.artifact }}
+
+      - name: Upload the Buzz provider
+        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
+        with:
+          name: ${{ matrix.provider_artifact }}
+          path: dist/${{ matrix.provider_artifact }}
 
   openapi:
     name: Export OpenAPI spec
@@ -506,7 +548,7 @@ jobs:
         uses: actions/download-artifact@3e5f45b2cfb9172054b4087a40e8e0b5a5461e7c # v8.0.1
         with:
           path: artifacts
-          pattern: "{fountain-*,openapi.json}"
+          pattern: "{fountain-*,buzz-backend-fountain-*,openapi.json}"
 
       - name: Create release
         uses: softprops/action-gh-release@efb35369e0ad2afab669f228072c1b0d510eae64 # v3.0.3
@@ -519,6 +561,10 @@ jobs:
             artifacts/fountain-linux-arm64/fountain-linux-arm64
             artifacts/fountain-darwin-amd64/fountain-darwin-amd64
             artifacts/fountain-darwin-arm64/fountain-darwin-arm64
+            artifacts/buzz-backend-fountain-linux-amd64/buzz-backend-fountain-linux-amd64
+            artifacts/buzz-backend-fountain-linux-arm64/buzz-backend-fountain-linux-arm64
+            artifacts/buzz-backend-fountain-darwin-amd64/buzz-backend-fountain-darwin-amd64
+            artifacts/buzz-backend-fountain-darwin-arm64/buzz-backend-fountain-darwin-arm64
             artifacts/openapi.json/openapi.json
           fail_on_unmatched_files: true
