+++ cli/credentials/credentials_test.go
