+++ cli/internal/cmd/apply.go
@@ -6,7 +6,7 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/BinaryBourbon/fountain/cli/internal/api"
+	"github.com/BinaryBourbon/fountain/cli/api"
 	"github.com/BinaryBourbon/fountain/cli/internal/manifest"
 	"github.com/BinaryBourbon/fountain/cli/internal/output"
 	"github.com/BinaryBourbon/fountain/cli/internal/secrets"
