+++ cli/internal/cmd/auth.go
@@ -14,9 +14,9 @@ import (
 	"strings"
 	"time"
 
-	"github.com/BinaryBourbon/fountain/cli/internal/api"
-	"github.com/BinaryBourbon/fountain/cli/internal/config"
-	"github.com/BinaryBourbon/fountain/cli/internal/credentials"
+	"github.com/BinaryBourbon/fountain/cli/api"
+	"github.com/BinaryBourbon/fountain/cli/config"
+	"github.com/BinaryBourbon/fountain/cli/credentials"
 	"github.com/spf13/cobra"
 	"golang.org/x/term"
 )
