+++ cli/internal/cmd/clientopts.go
@@ -1,8 +1,8 @@
 package cmd
 
 import (
-	"github.com/BinaryBourbon/fountain/cli/internal/api"
-	"github.com/BinaryBourbon/fountain/cli/internal/credentials"
+	"github.com/BinaryBourbon/fountain/cli/api"
+	"github.com/BinaryBourbon/fountain/cli/credentials"
 )
 
 // activeOpts builds credentials.Opts from the parsed --profile flag.
