+++ cli/internal/cmd/conv.go
@@ -13,7 +13,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/BinaryBourbon/fountain/cli/internal/api"
+	"github.com/BinaryBourbon/fountain/cli/api"
 	"github.com/BinaryBourbon/fountain/cli/internal/output"
 	"github.com/BinaryBourbon/fountain/cli/internal/sse"
 	"github.com/spf13/cobra"
