+++ cli/internal/cmd/keys.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/BinaryBourbon/fountain/cli/internal/api"
+	"github.com/BinaryBourbon/fountain/cli/api"
 	"github.com/BinaryBourbon/fountain/cli/internal/output"
 	"github.com/spf13/cobra"
 )
